training/requirements.txt
# Python Dependencies for ISL Model Training

# Core ML/DL frameworks
tensorflow>=2.13.0
numpy>=1.24.0
Pillow>=10.0.0

# Kaggle API for dataset download
kaggle>=1.5.16

# Data processing
pandas>=2.0.0
scikit-learn>=1.3.0

# Visualization (optional, for analysis)
matplotlib>=3.7.0
seaborn>=0.12.0

# Progress bars
tqdm>=4.65.0

# Jupyter notebook support (optional)
jupyter>=1.0.0
ipykernel>=6.25.0

# Model evaluation
scikit-image>=0.21.0

# Additional utilities
opencv-python>=4.8.0
